+++ .github/workflows/rust.yml
@@ -12,6 +12,8 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
+      - name: Set env
+        run: echo "WASM_TAG=index-v0-1-0" >> $GITHUB_ENV
       - uses: actions-rust-lang/setup-rust-toolchain@v1
         with:
           toolchain: beta
