+++ Cargo.toml
@@ -2,7 +2,6 @@
 resolver = "2"
 members = [
     "client",
-    "consumer",
     "server",
     "shared",
     "state"
@@ -13,12 +12,12 @@ members = [
 horfimbor-eventsource = "0.3"
 anyhow = "1.0"
 clap = { version = "4.4", features = ["derive"] }
-redis= { version = "0.26", features = ["tokio-rustls-comp"] }
+redis= { version = "0.28", features = ["tokio-rustls-comp"] }
 rocket = { version = "0.5", features = ["json"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1"
-thiserror = "1.0"
-eventstore = "3.0"
+thiserror = "2.0"
+eventstore = "4.0"
 uuid = { version = "1.1", features = ["v4", "serde"] }
 dotenvy = "0.15"
