+++ Dockerfile
@@ -7,7 +7,7 @@
 ################################################################################
 # Create a stage for building the application.
 
-ARG RUST_VERSION=1.75.0
+ARG RUST_VERSION=1.84.0
 FROM rust:${RUST_VERSION}-slim-bookworm AS build
 WORKDIR /app
 
@@ -29,7 +29,6 @@ ENV WASM_TAG=${WASM_TAG}
 RUN --mount=type=bind,source=Cargo.toml,target=Cargo.toml \
     --mount=type=bind,source=Cargo.lock,target=Cargo.lock \
     --mount=type=bind,source=client,target=client \
-    --mount=type=bind,source=consumer,target=consumer \
     --mount=type=bind,source=server,target=server \
     --mount=type=bind,source=shared,target=shared \
     --mount=type=bind,source=state,target=state \
@@ -39,9 +38,8 @@ RUN --mount=type=bind,source=Cargo.toml,target=Cargo.toml \
 #!/bin/bash
 
 set -e
-cargo build --locked --release --bin template-server --bin template-consumer
+cargo build --locked --release --bin template-server
 cp ./target/release/template-server /bin/template-server
-cp ./target/release/template-consumer /bin/template-consumer
 
 WASM_TAG=${WASM_TAG//./-}
 
@@ -49,43 +47,6 @@ wasm-pack build ./client --target web --out-dir /bin/webcomponent --out-name ind
 
 EOF
 
-
-################################################################################
-# Create a new stage for running the application that contains the minimal
-# runtime dependencies for the application. This often uses a different base
-# image from the build stage where the necessary files are copied from the build
-# stage.
-#
-# The example below uses the debian bookworm image as the foundation for running the app.
-# By specifying the "bookworm-slim" tag, it will also use whatever happens to be the
-# most recent version of that tag when you build your Dockerfile. If
-# reproducability is important, consider using a digest
-# (e.g., debian@sha256:ac707220fbd7b67fc19b112cee8170b41a9e97f703f588b2cdbbcdcecdd8af57).
-FROM debian:bookworm-slim AS final-consumer
-
-# add certificate
-RUN apt update \
-    && apt install -y ca-certificates libsasl2-dev libssl-dev
-RUN update-ca-certificates
-
-# Create a non-privileged user that the app will run under.
-# See https://docs.docker.com/develop/develop-images/dockerfile_best-practices/#user
-ARG UID=10001
-RUN adduser \
-    --disabled-password \
-    --gecos "" \
-    --home "/nonexistent" \
-    --shell "/sbin/nologin" \
-    --no-create-home \
-    --uid "${UID}" \
-    appuser
-USER appuser
-
-# Copy the executable from the "build" stage.
-COPY --from=build /bin/template-consumer /bin/
-
-# What the container should run when it is started.
-CMD ["/bin/template-consumer"]
 
 ################################################################################
 # Create a new stage for running the application that contains the minimal
