+++ justfile
@@ -1,5 +1,11 @@
+set shell := ["bash", "-uc"]
+set dotenv-load
+
 startup:
     docker compose up -d
+
+stop:
+    docker compose down
 
 watch-client:
     cargo watch -w client -w shared -- \
@@ -10,12 +16,9 @@ watch-client:
 
 watch-server:
     cargo watch -w server -w shared -w state -i server/web/ -i server/templates \
-        -x "run -p template-server"
-
-watch-consume kind:
-    cargo watch -w consumer -w shared -w state \
-      -x "run -p template-consumer -- --consumer {{kind}}"
+        -x "run -p template-server service"
 
 precommit:
     cargo fmt
     cargo clippy -- -D clippy::expect_used -D clippy::panic  -D clippy::unwrap_used
+    cargo test
