+++ server/Cargo.toml
@@ -21,9 +21,11 @@ serde_json.workspace = true
 anyhow.workspace = true
 chrono = "0.4"
 dotenvy.workspace = true
+tokio = "1.43"
+
 
 [dependencies.rocket_dyn_templates]
-version = "0.1.0-rc.3"
+version = "0.2"
 features = ["tera"]
 
 [build-dependencies]
