+++ server/src/consumer/dto.rs
@@ -1,9 +1,11 @@
-use crate::{TemplateDtoCache, TemplateDtoRepository, GROUP_NAME, STREAM_NAME};
+use crate::{TemplateDtoCache, TemplateDtoRepository, STREAM_NAME};
 use anyhow::Result;
 use eventstore::Client;
 use horfimbor_eventsource::repository::Repository;
 use horfimbor_eventsource::Stream;
 use redis::Client as Redis;
+
+const GROUP_NAME: &str = "dto";
 
 pub async fn cache_dto(redis_client: Redis, event_store_db: Client) -> Result<()> {
     let dto_redis = TemplateDtoCache::new(redis_client.clone());
