+++ server/src/consumer/state.rs
@@ -1,9 +1,11 @@
-use crate::{TemplateRepository, TemplateStateCache, GROUP_NAME, STREAM_NAME};
+use crate::{TemplateRepository, TemplateStateCache, STREAM_NAME};
 use anyhow::Result;
 use eventstore::Client;
 use horfimbor_eventsource::repository::Repository;
 use horfimbor_eventsource::Stream;
 use redis::Client as Redis;
+
+const GROUP_NAME: &str = "state";
 
 pub async fn cache_state(redis_client: Redis, event_store_db: Client) -> Result<()> {
     let dto_redis = TemplateStateCache::new(redis_client.clone());
