+++ state/Cargo.toml
@@ -15,7 +15,7 @@ template-shared = { path = "../shared", features = ["server"] }
 [dev-dependencies]
 tokio = { version = "1.21", features = ["rt", "rt-multi-thread", "macros"] }
 async-trait = "0.1"
-cucumber = { version = "0.20.0" }
+cucumber = { version = "0.21.1" }
 anyhow.workspace = true
 
 [[test]]
